Attempting to fix intermediate result/loop issue
Currently, it appears our fix is too eager. We probably have to add
something extra to assignments where we add instead of overwrite to the
list of possible values when we're in a loop (to accomodate the first
iteration).

No, wait, actually, that's not precisely the case; instead, we have to,
at the start of analysing the second loop, make sure the table is a
merge of both the pre-iter state and post-iter state.


Former-commit-id: f404557d144b5290dd86b1c9145bd86cb66bd86a