Fixed some bugs for the tutorial
TODO: Fix the instance planner to deal with loops.

TODO: Fix that Brane too-eagerly merges folders iff it's reading the
same result as it's writing to. Working on a workaround package
(`copy_result`, see [brane-std](https://github.com/epi-project/brane-std))
until we fix it, since this defs seems to be a larger issue with result
naming...

Former-commit-id: 473802d663dc55292057b683651429ebc9beeec8