Adding headers and boilerplate for VirtualMachines
We also gave a stab at moving the virtual machines to `brane-tsk`, but
in hindsight, this doesn't really make sense. Instead, let's just rename
the `brane-tsk-c` in `brane-cli-c` and depend on `brane-cli`, with the
philosophy that this `brane-ide` is indeed another kind of `cli`-like
client.

Man we desperately need to sit down and finish the new syntax, then we
can properly rearrange everything...