chore(deps): Add dependabot
Unfortunately, as of now, it is impossible to validate dependabot
configurations beforehand, I checked it twice, but there might be
something I missed. Right now, the limit for open dependabot pull
requests for version updates is set to five at a time. This might be a
little on the low end for a project of this size, but we shall see.